requirements.txt
pyTelegramBotAPI==4.20.0
APScheduler==3.10.4
SQLAlchemy==2.0.31
python-dotenv==1.0.1